requirements.txt
absl-py==0.10.0
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.3.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cloudpickle==1.6.0
cycler==0.11.0
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.3
entrypoints==0.3
Flask==1.1.2
flatbuffers==1.12
future==0.18.2
gast==0.3.3
google-auth==1.27.1
google-auth-oauthlib==0.4.3
google-pasta==0.2.0
googleapis-common-protos==1.53.0
graphsurgeon @ file:///workspace/TensorRT-7.2.2.3/graphsurgeon/graphsurgeon-0.4.5-py2.py3-none-any.whl
grpcio==1.32.0
h5py==2.10.0
horovod @ file:///opt/tensorflow/horovod-source/dist/horovod-0.21.3.tar.gz
hyperopt==0.2.5
idna==2.10
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.1.0
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyter-tensorboard @ git+https://github.com/cliffwoolley/jupyter_tensorboard.git@ffa7e26138b82549453306e06b535a9ac36db17a
jupyterlab==2.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
jupytext==1.10.2
keras==2.7.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
llvmlite==0.37.0
Markdown==3.3.4
markdown-it-py==0.6.2
MarkupSafe==1.1.1
matplotlib==3.4.3
mdit-py-plugins==0.2.5
mistune==0.8.4
mock==3.0.5
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
networkx==2.6.3
nltk==3.4.5
notebook==6.2.0
numba==0.54.1
numpy==1.19.4
nvidia-dali-cuda110==0.31.0
nvidia-dali-tf-plugin-cuda110==0.31.0
nvidia-tensorboard-plugin-dlprof @ file:///nvidia/opt/tensorboard_install/nvidia_tensorboard_plugin_dlprof-0.12-py3-none-any.whl
nvtx-plugins==0.1.8
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.9
pandas==1.3.4
pandocfilters==1.4.3
parso==0.8.1
patsy==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==8.4.0
polygraphy==0.25.0
portpicker==1.3.1
prometheus-client==0.9.0
promise==2.3
prompt-toolkit==3.0.16
protobuf==3.15.4
psutil==5.7.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.8.0
pyparsing==2.4.7
pyprg==0.1.1b7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.3
PyYAML==5.4.1
pyzmq==22.0.3
regex==2021.11.2
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==1.0.1
scipy==1.4.1
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
sktime==0.8.1
statsmodels==0.12.1
tensorboard @ file:///nvidia/opt/tensorboard_install/tensorboard-2.4.1-py3-none-any.whl
tensorboard-plugin-wit==1.8.0
tensorflow @ file:///tmp/pip/tensorflow-2.4.0%2Bnv-cp38-cp38-linux_x86_64.whl
tensorflow-addons @ file:///opt/tensorflow/tf-addons/artifacts/tensorflow_addons-0.11.2-cp38-cp38-linux_x86_64.whl
tensorflow-datasets==3.2.1
tensorflow-estimator==2.4.0
tensorflow-metadata==0.28.0
tensorrt @ file:///workspace/TensorRT-7.2.2.3/python/tensorrt-7.2.2.3-cp38-none-linux_x86_64.whl
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
threadpoolctl==3.0.0
toml==0.10.2
tornado==6.1
tqdm==4.59.0
traitlets==5.0.5
typeguard==2.11.1
typing-extensions==3.7.4.3
uff @ file:///workspace/TensorRT-7.2.2.3/uff/uff-0.6.9-py2.py3-none-any.whl
urllib3==1.26.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
